+++ src/server/models/schemas.py
@@ -1,7 +1,7 @@
 # models/schemas.py
 from pydantic import BaseModel, field_validator
 from typing import List
-from config.constants import SUPPORTED_LANGUAGES
+
 
 
 class SynthesizeRequest(BaseModel):
