+++ src/server/routes/health.py
