+++ src/server/utils/device_utils.py
